move api to command dirs
