+++ src/api/mod.rs
@@ -1,6 +1,1 @@
-pub mod alerts;
 pub mod client;
-pub mod dashboards;
-pub mod dataprime;
-pub mod metrics;
-pub mod semantic_search;
