+++ src/commands/dataprime/api.rs
@@ -6,7 +6,7 @@ use serde_json::{json, Value};
 use crate::error::Result;
 use crate::Tier;
 
-use super::client::CxClient;
+use crate::api::client::CxClient;
 
 // ── Log types ─────────────────────────────────────────────────────────────────
